backend/requirements.txt
# ── Web Framework ─────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0

# ── AI ────────────────────────────────────────────────────────────────────
openai==1.30.1

# ── HTTP Client (for Airtable, Slack calls) ───────────────────────────────
httpx==0.27.0

# ── Data ──────────────────────────────────────────────────────────────────
pandas==2.2.2
faker==25.0.0

# ── Utilities ─────────────────────────────────────────────────────────────
python-dotenv==1.0.1
aiofiles==23.2.1